+++ src/Pagination/PaginationManipulator.php
@@ -4,7 +4,10 @@ namespace Nuwave\Lighthouse\Pagination;
 
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use GraphQL\Language\AST\TypeNode;
 use GraphQL\Language\Parser;
+use Illuminate\Container\Container;
+use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
@@ -127,7 +130,7 @@ after: String
 GRAPHQL
         );
 
-        $fieldDefinition->type = Parser::typeReference(/** @lang GraphQL */ "{$connectionTypeName}!");
+        $fieldDefinition->type = $this->paginationResultType($connectionTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
@@ -189,7 +192,7 @@ page: Int
 GRAPHQL
 );
 
-        $fieldDefinition->type = Parser::typeReference(/** @lang GraphQL */"{$paginatorTypeName}!");
+        $fieldDefinition->type = $this->paginationResultType($paginatorTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
@@ -225,7 +228,7 @@ page: Int
 GRAPHQL
         );
 
-        $fieldDefinition->type = Parser::typeReference(/** @lang GraphQL */"{$paginatorTypeName}!");
+        $fieldDefinition->type = $this->paginationResultType($paginatorTypeName);
         $parentType->fields = ASTHelper::mergeUniqueNodeList($parentType->fields, [$fieldDefinition], true);
     }
 
@@ -247,5 +250,26 @@ GRAPHQL
             );
 
         return $description.$definition;
+    }
+
+    /**
+     * @return \GraphQL\Language\AST\NamedTypeNode|\GraphQL\Language\AST\NonNullTypeNode
+     */
+    protected function paginationResultType(string $typeName): TypeNode
+    {
+        /** @var \Illuminate\Contracts\Config\Repository $config */
+        $config = Container::getInstance()->make(ConfigRepository::class);
+        $nonNull = $config->get('lighthouse.non_null_pagination_results')
+            ? '!'
+            : '';
+
+        /**
+         * We do not wrap the typename in [], so this will never be a ListOfTypeNode.
+         *
+         * @var \GraphQL\Language\AST\NamedTypeNode|\GraphQL\Language\AST\NonNullTypeNode $nonNullTypeNode
+         */
+        $nonNullTypeNode = Parser::typeReference(/** @lang GraphQL */ "{$typeName}{$nonNull}");
+
+        return $nonNullTypeNode;
     }
 }