+++ CHANGELOG.md
@@ -13,8 +13,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Changed
 
-- Add config setting `non_null_pagination_results` to mark the generated result type of paginated lists as non-nullable
-- Improve descriptions of generated pagination types
+- Add config setting `non_null_pagination_results` to mark the generated result type of paginated lists as non-nullable https://github.com/nuwave/lighthouse/pull/1878
+- Improve descriptions of generated pagination types https://github.com/nuwave/lighthouse/pull/1878
 
 ## v5.12.7